Fixing The Translate Function!
When a user moved the work history right or left with a touch action, the start and end variables were not properly being reset neither were they being checked
if the variable was equal to null/undefined nor if it was reset and no touch event was fired.

This cause an unexpected side effect where if a user moved the slider left or right and then tapped on it the move event would not fire leaving it as the
same value as before and causing the slider to move when not prompted to.

A few styles were also updated for better responsiveness.